Add design system, accessibility, error-handling, and structural cleanup
Cleanliness/scalability pass across the app:

Foundation
- DesignSystem/DesignTokens.swift: Theme tokens (spacing, radius, icon
  sizes, shadows, animation, typography) + .pressable()/.shadow() helpers
- DesignSystem/Haptics.swift: semantic @MainActor haptics API
- Components/StateViews.swift: reusable Loading/Error/Empty state views
- cardStyle()/prominentButton() now use Theme tokens

Accessibility
- VoiceOver labels + coherent elements on RecipeCardView, TagView,
  ProfileAvatar, AddRecipeMenu, and the cook-mode timer (spoken time value)
- Consolidated scattered UIImpactFeedbackGenerator usage into Haptics
- Dynamic Type (@ScaledMetric) on the timer readout

Error handling
- bestEffort(_:logger:) helper; converted 14 silent `try? await` write
  failures into logged best-effort calls (no behavior change, now diagnosable)

Structural
- Split CollectionDetailView 1850->1090 (CollectionRecipeSelectorSheet,
  ConformanceFixSheet)
- Split FriendsTabView 1186->586 (FriendsTabComponents, AllFriendsListViews,
  InviteFriendsSheetView)

Tests
- SearchTabViewModelTests (7) and UserProfileViewModelTests (6), all passing

Design
- Typography scale + PressableScaleStyle; applied tokens to RecipeCardView
