+++ Cauldron/ContentView.swift
@@ -441,7 +441,9 @@ struct ContentView: View {
             if let cloudUsers = try? await userCloudService.fetchUsers(byUserIds: Array(relatedUserIds)) {
                 for cloudUser in cloudUsers {
                     usersById[cloudUser.id] = cloudUser
-                    try? await sharingRepository.save(cloudUser)
+                    await bestEffort("Cache related user") {
+                        try await sharingRepository.save(cloudUser)
+                    }
                 }
             }
